Split big cli.go into small files && change initialization step with utils one in some of them
